perf: read file worker
