+++ packages/service/worker/file/extension/pptx.ts
@@ -1,11 +1,11 @@
-import { ReadFileByBufferParams, ReadFileResponse } from './type.d';
+import { ReadRawTextByBuffer, ReadFileResponse } from '../type';
 // import { parseOfficeAsync } from 'officeparser';
-import { parseOffice } from './parseOffice';
+import { parseOffice } from '../parseOffice';
 
 export const readPptxRawText = async ({
   buffer,
   encoding
-}: ReadFileByBufferParams): Promise<ReadFileResponse> => {
+}: ReadRawTextByBuffer): Promise<ReadFileResponse> => {
   const result = await parseOffice({
     buffer,
     encoding: encoding as BufferEncoding,
